% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_api_version.R
\name{account_api_version}
\alias{account_api_version}
\title{Get the version of the API the account uses}
\usage{
account_api_version(account)
}
\arguments{
\item{account}{The account object.}
}
\value{
A string: either V3 or V4.
}
\description{
There are currently only two versions, V3 and V4. This library
supports only the V4 format. See the originaly brandseyer library
if you would like to read data from a V3 account.
}
