% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_brands.R
\name{account_brands}
\alias{account_brands}
\title{Fetch account brands}
\usage{
account_brands(account)
}
\arguments{
\item{account}{An account object.}
}
\value{
A tibble of brand information
}
\description{
Fetches brand information, returned as a tibble, for the given
account. It includes brand names, associated phrases, and so on.
}
\examples{

\dontrun{
# Fetch brand information
account("TEST01AA") \%>\%
  account_brands()

# Find deleted parent brands
account("TEST01AA") \%>\%
  account_brands() \%>\%
  dplyr::filter(is.na(parent), deleted)

# Fetch phrases without using account_phrases
account("TEST01AA") \%>\%
  account_brands() \%>\%
  dplyr::select(id, phrases) \%>\%
  dplyr::rename(brand.id = id) \%>\%
  tidyr::unnest(phrases) \%>\%
  dplyr::rename(phrase.id = id)
}
}
