% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_client_code.R
\name{account_client_code}
\alias{account_client_code}
\title{Get an account's client code}
\usage{
account_client_code(account)
}
\arguments{
\item{account}{The account object, or list of account objects.}
}
\value{
A string holding the account's client code.
}
\description{
Returns a vector of account client codes.
}
\examples{

account("TEST01AA") \%>\%
  account_client_code()

account("TEST01AA", "TEST02AA") \%>\%
  account_client_code()
}
