% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_manager.R
\name{account_manager}
\alias{account_manager}
\alias{account_manager.brandseyer2.account}
\title{See your account manager}
\usage{
account_manager(account)

\method{account_manager}{brandseyer2.account}(account)
}
\arguments{
\item{account}{An account object}
}
\value{
A list containing the name and email address of your account manager.
}
\description{
This provides the name and email address of your
account manager. You can contact this person if you have any questions
concerning your account.
}
\details{
The returned object provides two fields:

\enumerate{
  \item \code{$name} Holds your account manager's name
  \item \code{$email} Holds your account manager's email address
}
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Your manager for a particular account.
}}

\examples{
\dontrun{
ac <- account("TEST01AA")
manager <- account_manager(ac)

manager$name
manager$email
}

}
