% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_name.R
\name{account_name}
\alias{account_name}
\title{Get an account's name}
\usage{
account_name(account)
}
\arguments{
\item{account}{The account object.}
}
\value{
A string holding the account's name
}
\description{
Returns a vector of names for the accounts passed as the argument.
}
\examples{

account("TEST01AA") \%>\%
  account_name()

account("TEST01AA", "TEST02AA") \%>\%
  account_name()
}
