% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_tags.R
\name{account_tags}
\alias{account_tags}
\title{Fetch tags for an account}
\usage{
account_tags(account)
}
\arguments{
\item{account}{An account object.}
}
\value{
A tibble of tag information. Includes the name, namespace, and description of the
tags. Note that topics are stored in the 'topics' namespace.
}
\description{
Fetches tag information information, returned as a tibble, for the given
account.
}
\examples{
\dontrun{

# See what namespaces are in your account
account("TEST01AA") \%>\%
  account_tags() \%>\%
  dplyr::select(namespace) \%>\%
  table()

# Find topics
account("TEST01AA") \%>\%
  account_tags() \%>\%
  dplyr::filter(namespace == "topic")
}

}
