% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_timezone.R
\name{account_timezone}
\alias{account_timezone}
\title{Get an account's timezone}
\usage{
account_timezone(account)
}
\arguments{
\item{account}{The account object.}
}
\value{
A string holding the timezone that account data is reported in.
}
\description{
An account's data is reported in a particular timezone, set for that account.
All dates provided in filters are assumed to be given in that timezone.
}
\author{
Constance Neeser
}
