% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_topics.R
\name{account_topics}
\alias{account_topics}
\alias{account_topics.brandseyer2.account}
\alias{account_topics.list}
\title{Fetch topics for an account}
\usage{
account_topics(accounts)

\method{account_topics}{brandseyer2.account}(accounts)

\method{account_topics}{list}(accounts)
}
\arguments{
\item{accounts}{One or more account objects.}
}
\value{
A tibble of topic information. Includes the name and topic description.
}
\description{
Fetches the topics that are used in an account.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Get topics for a single account.

\item \code{list}: Get topics for a list of accounts
}}

