% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_type.R
\name{account_type}
\alias{account_type}
\title{Get an account's type}
\usage{
account_type(account)
}
\arguments{
\item{account}{The account object.}
}
\value{
A string indicating the account's type, such as an internal account, paid account, etc.
}
\description{
Get an account's type
}
