% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_add_pickedup.R
\name{add_pickedup}
\alias{add_pickedup}
\title{Add a pickedup date restriction to a filter.}
\usage{
add_pickedup(filter, timezone, time = Sys.time())
}
\arguments{
\item{filter}{The filter to amend.}

\item{timezone}{The timezone that the time is given in}

\item{time}{An optional time to use. Otherwise uses the current time.}
}
\value{
The ammended filter.
}
\description{
Add a pickedup date restriction to a filter.
}
