% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{am_i_brandseye}
\alias{am_i_brandseye}
\title{Are you authorised as a BrandsEye staff member?}
\usage{
am_i_brandseye()
}
\value{
A logical value.
}
\description{
Returns \code{TRUE} if and only if you are a BrandsEye staff.
}
