% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate yourself with BrandsEye}
\usage{
authenticate(key, save = FALSE)
}
\arguments{
\item{key}{Your API key. If you don't know this, your account manager can give you one.}

\item{save}{Set to true if you would like this key to be saved to disc and loaded
automatically the next time that you use this library.}
}
\value{
Your authentication details
}
\description{
Provides a means to authenticate yourself with the BrandsEye API. This is done
using an API key that your account manager representative has provided for you.
}
\details{
All authentication with the BrandsEye API is done via a special, user specific
key. This key can be provided to you by your account manager.

It's possible to also have your key saved to disc and automatically loaded
for you when you next use the library. Set the \code{save} parameter to \code{TRUE}
in order to do this.

If you would like to see who you are currently logged in as, see the
\code{\link[=whoami]{whoami()}} function.
}
\examples{
\dontrun{

authenticate("adfd42345f534fgdfgd")

}

}
\seealso{
\code{\link[=account_manager]{account_manager()}}
}
\author{
Constance Neeser
}
