% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authentication_filename}
\alias{authentication_filename}
\title{Returns the file name in which you can save your authentication information.}
\usage{
authentication_filename()
}
\description{
Returns the file name in which you can save your authentication information.
}
