% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_errors.R
\name{check_errors}
\alias{check_errors}
\title{Checking for errors}
\usage{
check_errors(data)
}
\arguments{
\item{data}{Data returned from an \code{httr} call.}
}
\description{
Pass the results from httr calls to the API to examine errors
and report on them in a standard way.
}
\author{
Constance Neeser
}
