% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-confirm.R
\name{confirm}
\alias{confirm}
\title{Confirm an action.}
\usage{
confirm(message)
}
\arguments{
\item{message}{A message to display to the user.}
}
\description{
Asks the user to answer Yes or No after prompting them with a message.
}
\author{
Constance Neeser
}
