% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_add_tags.R
\name{create_tags}
\alias{create_tags}
\title{Add tags to an account}
\usage{
create_tags(x, name, description)
}
\arguments{
\item{x}{An account to add the tag to.}

\item{name}{A vector of tag names.}

\item{description}{An optional description for each tag. If given, must be
the same length as \code{name}.}
}
\value{
A tibble containing the new tag's details, including the tag's ID.
}
\description{
This will add a new tag to an account. This tag will be added
in the tag namespace. If a tag with the same name already exists,
this will do nothing. If a tag with the same name exists, and is deleted,
it will create a new tag with a new ID.
}
\examples{

\dontrun{

# Create one tag
account("TEST01AA") \%>\%
  create_tags("tag1")

# Create a tag with a description
account("TEST01AA") \%>\%
  create_tags("tag1", "A description")

# Create multiple tags
account("TEST01AA") \%>\%
  create_tags(c("tag1", "tag2"))

# Create multiple tags, with descriptions
account("TEST01AA") \%>\%
  create_tags(c("tag1", "tag2"),
              c("description 1", "description2"))

}
}
