% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_dashboards.R
\name{dashboards}
\alias{dashboards}
\alias{dashboards.brandseyer2.account}
\title{List the dashboards in an account}
\usage{
dashboards(x, d, ...)

\method{dashboards}{brandseyer2.account}(x, d, section, ...)
}
\arguments{
\item{x}{An account object to query.}

\item{d}{An optional value to filter the dashboards on. Can be a vector
of integer IDs for the wanted dashboards, or a vector of
characters giving words to perform partial string matches
on dashboard names with.}

\item{...}{Additional parameters for methods}

\item{section}{An optional numeric vector of section IDs, or
character vector to match against section titles. This specifies
which sections to return, similarly to calling \code{\link[=sections]{sections()}}. You
can use the special string \code{'.'} to select all sections.}
}
\value{
A tibble of dashboard data.
}
\description{
In \href{https://analyse.brandseye.com}{Analyse}, BrandsEye's analysis tool,
people can set up dashboards to query our data. This function exposese those
dashboards as a tibble, allowing you to query sections, metrics and their
associated filters.
}
\details{
Dashboards can be filtered in various ways. The second argument, \code{d}, can be
either an integer or list of integers, specifying the particular dashboard
IDs that are wanted. The second argument can also be a character vector,
to give partial matches against dashboard names. Case is ignored.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Read dashboard data for an \code{\link[=account]{account()}} object.
}}

\examples{

\dontrun{

# Get all dashboards.
account("TEST01AA") \%>\%
  dashboards()

# Get the dashboard with ID 1.
account("TEST01AA") \%>\%
  dashboards(1)

# Get dashboards 1 and 21.
account("TEST01AA") \%>\%
  dashboards(c(1, 21))

# Get dashboards whose name contains the string 'overview'.
account("TEST01AA") \%>\%
  dashboards("overview")

# Get dashboards whose name contains the string 'overview' or the string 'who'.
account("TEST01AA") \%>\%
  dashboards(c("overview", "who"))

}
}
\seealso{
\code{\link[=sections]{sections()}} to pull section information for a dashboard.

\code{\link[=metrics]{metrics()}} to pull metric information for sections.
}
