% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{data_models}
\alias{data_models}
\alias{data_model_countries}
\alias{data_model_categories}
\alias{data_model_languages}
\alias{data_model_networks}
\title{Explore the API data models}
\usage{
data_model_countries()

data_model_categories()

data_model_languages()

data_model_networks()
}
\description{
Provides a table of values and the IDs that represent
them in the filter language. For instance, country names and the
codes used to represent them.
}
\details{
\code{\link[=data_model_countries]{data_model_countries()}} supplies a tibble
of country names and their ISO 3166-1 Alpha 2 country
codes, as used in the filter language.

Our media types are listed in \code{\link[=data_model_categories]{data_model_categories()}}.
\code{ENTERPRISE} represents mentions coming from the brand itself
(i.e., the brand that the mention is matched against), while
\code{CONSUMER} are most other people. \code{PRESS} represents media sources.

We use 2-letter ISO 639-1 codes to represent
our languages, which are listed using \code{\link[=data_model_languages]{data_model_languages()}}.

The \code{socialNetwork} field in the
filter language takes IDs provided by \code{\link[=data_model_networks]{data_model_networks()}}.
}
