% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_brand.R
\name{filter_brand}
\alias{filter_brand}
\title{Describe a brand}
\usage{
filter_brand(id, code = NULL, name = NULL)
}
\arguments{
\item{id}{The brand ID}

\item{code}{An account code (optional)}

\item{name}{The brand's name (optional)}
}
\value{
A \code{\link[=filter_brand]{filter_brand()}} object.
}
\description{
Describes a brand, and optionally, its account and name.
}
\examples{

filter_brand(1)

filter_brand(1, "TEST01AA")

filter_brand(1, "TEST01AA", "The first brand")
}
\seealso{
This is used by \code{\link[=query]{query()}}.
}
