% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_language.R
\name{filter_mentions}
\alias{filter_mentions}
\title{Filter for only specific mentions}
\usage{
filter_mentions(.account, filter)
}
\arguments{
\item{.account}{An account or \code{\link[=query]{query()}} object.}

\item{filter}{A filter string.}
}
\value{
A query object
}
\description{
\code{filter_mentions()} causes counting and mention operations
to occur on only those mentions matching the filter. This is
part of the \code{\link[=query]{query()}} language.
}
\section{Data models}{


The filter language often uses IDs to represent countries, languages, and so on.
These can be found by using either the filter panel in \href{https://analyse.brandseye.com}{Analyse},
or the data model functions.
\itemize{
\item \code{\link[=data_model_countries]{data_model_countries()}}
\item \code{\link[=data_model_languages]{data_model_languages()}}
\item \code{\link[=data_model_currencies]{data_model_currencies()}}
\item \code{\link[=data_model_networks]{data_model_networks()}}
}
}

\examples{

account("TEST01AA") \%>\%
  filter_mentions("published inthelast week")
}
\seealso{
Other verbs for the query language: \code{\link[=group_mentions_by]{group_mentions_by()}},
\code{\link[=compare_mentions]{compare_mentions()}}, \code{\link[=with_mention_fields]{with_mention_fields()}}, \code{\link[=with_mention_order]{with_mention_order()}}, \code{\link[=with_account]{with_account()}}.

\code{\link[=query]{query()}} is a way to manually create queries.
}
