% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_query.R
\name{get_query_accounts}
\alias{get_query_accounts}
\title{See the accounts in a query.}
\usage{
get_query_accounts(query)
}
\arguments{
\item{query}{A \code{\link[=query]{query()}} object.}
}
\value{
A vector of account codes.
}
\description{
Returns a vector of account codes in a query. Ideal for
printing all accounts being queried in large queries.
}
\examples{

q <- account("TEST01AA") \%>\%
  filter_mentions("published inthelast week")
get_query_accounts(q)

}
