% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_query.R
\name{get_query_brands}
\alias{get_query_brands}
\title{See brands in a query}
\usage{
get_query_brands(query)
}
\arguments{
\item{query}{A \code{\link[=query]{query()}} object.}
}
\value{
A list of brands.
}
\description{
Returns a list of brands in a query. It is useful for
interactively displaying all the brands in a large query.
}
\examples{

q <- account("TEST01AA") \%>\%
  filter_mentions("published in thelast week")
get_query_brands(q)

}
