% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_language.R
\name{group_mentions_by}
\alias{group_mentions_by}
\title{Group mentions by shared criteria}
\usage{
group_mentions_by(.account, ..., .envir)
}
\arguments{
\item{.account}{An account or \code{\link[=query]{query()}} object.}

\item{...}{A list of fields to group by.}

\item{.envir}{An optional environment to perform substitutions in.}
}
\value{
A \code{\link[=query]{query()}} object.
}
\description{
\code{\link[=group_mentions_by]{group_mentions_by()}} allows you to specify how to group
mentions when counting them, such as by day, sentiment, etc.
This is part of the \code{\link[=query]{query()}} language.
}
\examples{

account("TEST01AA", "TEST02AA") \%>\%
  filter_mentions("published inthelast week") \%>\%
  group_mentions_by(published, sentiment)
}
\seealso{
Other verbs for the query language: \code{\link[=group_mentions_by]{group_mentions_by()}},
\code{\link[=compare_mentions]{compare_mentions()}}, \code{\link[=with_fields]{with_fields()}}, \code{\link[=with_mention_order]{with_mention_order()}}, \code{\link[=with_account]{with_account()}}.

\code{\link[=query]{query()}} is a way to manually create queries.
}
