% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_is_account.R
\name{is.account}
\alias{is.account}
\title{See if an object is a BrandsEyeR account object.}
\usage{
is.account(x)
}
\arguments{
\item{x}{A possible BrandsEyeR account object.}
}
\value{
TRUE if this is an account object, otherwise false.
}
\description{
See if an object is a BrandsEyeR account object.
}
