% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_is_account.R
\name{is_account}
\alias{is_account}
\title{See if an object is a account object}
\usage{
is_account(x)
}
\arguments{
\item{x}{A possible \code{\link[=account]{account()}} object.}
}
\value{
TRUE if this is an account object, otherwise false.
}
\description{
Returns true if and only if the argument is an \code{\link[=account]{account()}} object.
}
