% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_is_query.R
\name{is_query}
\alias{is_query}
\title{See if an object is a BrandsEyeR query}
\usage{
is_query(x)
}
\arguments{
\item{x}{A possible \code{\link[=query]{query()}} account object.}
}
\value{
TRUE if this is an account object, otherwise false.
}
\description{
Returns true if and only if the argument is a \code{\link[=query]{query()}} object.
}
