% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_mentions.R
\name{list_to_v4_mentions}
\alias{list_to_v4_mentions}
\title{Format list data as a tibble}
\usage{
list_to_v4_mentions(data, is_staff = am_i_brandseye())
}
\arguments{
\item{data}{List data, as possibly returned from \code{link{read_data}}().}

\item{is_staff}{A logical. TRUE iff the user is a staff member.}
}
\value{
A tibble of mention data.
}
\description{
This takes list data returned from the API and extracts
information for the mention table.
}
