% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_logs.R
\name{logs}
\alias{logs}
\alias{logs.brandseyer2.account}
\alias{logs.list}
\title{Read account logs}
\usage{
logs(x, from)

\method{logs}{brandseyer2.account}(x, from = (Sys.Date() \%m-\% months(1)))

\method{logs}{list}(x, from = (Sys.Date() \%m-\% months(1)))
}
\arguments{
\item{x}{An \code{\link[=account]{account()}} object or list of them.}

\item{from}{A date indicating from when to read dates.}
}
\value{
A tibble of log information, including the user that perform the action, and how many
times the action was performed.
}
\description{
Returns a tibble of logs for the account, in the last period stated.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: For \code{\link[=account]{account()}} objects.

\item \code{list}: For a list of accounts
}}

\examples{

\dontrun{
# Read logs for an account
account("TEST01AA") \%>\%
  logs()
}


\dontrun{
# Read for a list of accounts
account("TEST01AA", "TEST02AA") \%>\%
  logs()
}
}
