% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_logs_retrosent.R
\name{logs_retrosent}
\alias{logs_retrosent}
\alias{logs_retrosent.data.frame}
\alias{logs_retrosent.character}
\alias{logs_retrosent.brandseyer2.account}
\alias{logs_retrosent.list}
\title{Finding retrosends for accounts}
\usage{
logs_retrosent(x, ...)

\method{logs_retrosent}{data.frame}(x, ...)

\method{logs_retrosent}{character}(x, ...,
  .show.progress = interactive())

\method{logs_retrosent}{brandseyer2.account}(x, ...)

\method{logs_retrosent}{list}(x, ..., .show.progress = interactive())
}
\arguments{
\item{x}{A data frame of logs, or an account object (or list of account objects)}

\item{...}{Extra arguments}

\item{.show.progress}{When reading logs, whether to show a progress bar or not. See \code{\link[=logs]{logs()}}.}
}
\value{
A tibble giving retrosend information, such as the crowd, priority and filter used.
}
\description{
\code{\link[=logs_retrosent]{logs_retrosent()}} can be used to find retrosends that have occurred
for a particular \code{\link[=account]{account()}} or list of accounts. If you want
to specify a date range, use \code{\link[=logs]{logs()}} and pipe its results
through to \code{\link[=logs_retrosent]{logs_retrosent()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Reading from a data.frame or tibble of log information, such as that given
by \code{\link[=logs]{logs()}}.

\item \code{character}: From a vector of account codes.

\item \code{brandseyer2.account}: From an account object

\item \code{list}: From a list of account objects.
}}

\examples{

\dontrun{

account("TEST01AA") \%>\%
  logs_retrosent()

account_list() \%>\%
  logs() \%>\%
  logs_retrosent()

# Pull date for a particular time range
account("TEST01AA") \%>\%
  logs(Sys.Date() - lubridate::weeks(1)) \%>\%
  logs_retrosent()

# Pull data from a particular date up to today.
account("TEST01AA") \%>\%
  logs(lubridate::ymd("2018/08/01")) \%>\%
  logs_retrosent()
}
}
