% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_metrics.R
\name{metrics}
\alias{metrics}
\alias{metrics.data.frame}
\title{See metrics in a dashboard}
\usage{
metrics(x, m, ...)

\method{metrics}{data.frame}(x, m, collapse = TRUE, ...)
}
\arguments{
\item{x}{The tibble to pull widget information from.}

\item{m}{An optional filter for selecting metrics. A numeric
vector allows you to choose metrics with the given IDs. A
character vector allows you to choose metrics with matching captions.}

\item{...}{Parameters for other methods.}

\item{collapse}{A logical value indicating whether columns should be removed (collapsed)
or not.}
}
\value{
An unnested table of metrics.
}
\description{
This provides easy access to metric information stored in the metric
column of a tibble returned by \code{\link[=sections]{sections()}} and \code{\link[=dashboards]{dashboards()}}. It
essentially uses \code{\link[tidyr:unnest]{tidyr::unnest()}} to clean up the table.
}
\details{
This provides an additional field not available in the original metric
information called \code{filter}. This indicates the filter that the metric would have
used to call information from the BrandsEye API, and is a combination of the
section and metric filters.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Expand metric information for data frames.
}}

\seealso{
\code{\link[=dashboards]{dashboards()}} to pull dashboard information for an account.

\code{\link[=sections]{sections()}} to pull section information for an account.
}
