% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_phrases.R
\name{phrases}
\alias{phrases}
\alias{phrases.brandseyer2.account}
\alias{phrases.list}
\title{Fetch account phrases}
\usage{
phrases(accounts)

\method{phrases}{brandseyer2.account}(accounts)

\method{phrases}{list}(accounts)
}
\arguments{
\item{accounts}{One or more account objects.}
}
\value{
A tibble of phrase information.
}
\description{
Fetches phrase information, returned as a tibble, for the given
account. All of this information is included in \code{\link{brands}}
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Read phrases for only a single account

\item \code{list}: Create a table of phrases for the list of accounts given
}}

\examples{

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  phrases()
}
