% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_profiles.R
\name{profiles}
\alias{profiles}
\alias{profiles.character}
\alias{profiles.brandseyer2.account}
\alias{profiles.list}
\alias{profiles.data.frame}
\title{Read online profiles for an accounts}
\usage{
profiles(x, .show.progress)

\method{profiles}{character}(x, .show.progress = interactive())

\method{profiles}{brandseyer2.account}(x, .show.progress = interactive())

\method{profiles}{list}(x, .show.progress = interactive())

\method{profiles}{data.frame}(x, .show.progress = interactive())
}
\arguments{
\item{x}{An object to read profiles from.}

\item{.show.progress}{A logical indicating whether to show a progress bar or not.
By default, this only shows a progress bar in interactive environments.
Also, progress bars will only be shown if 3 or more accounts are
being examined.}
}
\value{
A tibble of log information, including the user that perform the action, and how many
times the action was performed.
}
\description{
Returns a tibble of online profiles for the account.
}
\details{
Online profiles can be read from various different kinds of object. An \code{\link[=account]{account()}}
object is the most basic, although profiles can also be read from a list of
account objects (in which case the returned tibble will include an account code column),
or from a tibble containing an account code column, such as that returned
by \code{\link[=account_list]{account_list()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Reach for an account code character vector.

\item \code{brandseyer2.account}: For \code{\link[=account]{account()}} objects.

\item \code{list}: For a list of accounts

\item \code{data.frame}: Read for a data frame, such as returned by \code{\link[=account_list]{account_list()}}
}}

\examples{

\dontrun{
# Read profiles for an account code
profiles("TEST01AA")
}


\dontrun{
# Read profiles for an account
account("TEST01AA") \%>\%
  profiles()
}


\dontrun{
# Read for a list of accounts
account("TEST01AA", "TEST02AA") \%>\%
  profiles()
}
\dontrun{
account_list() \%>\%
  profiles()
}
}
