% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_query.R
\name{query}
\alias{query}
\title{Create a query for mentions or counts}
\usage{
query(accounts = NULL, brands = NULL, timezones = NULL,
  filter = NULL, comparison = NULL, fields = NULL, grouping = NULL,
  ordering = NULL)
}
\arguments{
\item{accounts}{A character vector of accounts to use.}

\item{brands}{A list of brands to query against.}

\item{timezones}{The timezones to report dates in when counting mentions.}

\item{filter}{The filter to use when fetching or counting mentions.}

\item{comparison}{A list of subfilters to use when counting mentions.}

\item{fields}{Extra fields to return when fetching or counting mentions.}

\item{grouping}{How to group mentions when counting them.}

\item{ordering}{The order to return results in.}
}
\value{
A query object
}
\description{
\code{\link[=query]{query()}} creates a query to be used for fetching mentions, or
for counting them. You probably want to avoid using this directly,
and use one of the query verbs instead, such as \code{\link[=filter_mentions]{filter_mentions()}}.
\code{\link[=to_query]{to_query()}} is a convenient way to create a query object from
other things, such as an \code{\link[=account]{account()}}.
}
\details{
The query object is accepted by \code{\link[=count_mentions]{count_mentions()}} and \code{\link[=mentions]{mentions()}}.
The advantage of using query and its various verbs is that it handles
adding root brands to your filter for you, as well as filtering out
mentions using older versions of the API, or who have no brands altogether.

You can print out the brands in a query using \code{\link[=get_query_brands]{get_query_brands()}}.

Of special note, V4 accounts will not be added to a query.
}
\section{Query verbs}{


Various functions can be chained together using the pipe operator as 'verbs'
in a query sentence.
\itemize{
\item \code{\link[=filter_mentions]{filter_mentions()}} lets you choose what mentions to count or fetch.
\item \code{\link[=group_mentions_by]{group_mentions_by()}} lets you group mentions when counting.
\item \code{\link[=compare_mentions]{compare_mentions()}} lets you count and compare mentions from subfilters.
\item \code{\link[=with_fields]{with_fields()}} lets you select extra information fields, such as OTS, or engagement.
\item \code{\link[=with_order]{with_order()}} lets you order the results.
}
}

\examples{

query(accounts = c("TEST01AA", "TEST02AA"),
      filter = "published inthelast week",
      timezone = "UTC",
      brands = list(filter_brand(1), filter_brand(2))
)
}
\seealso{
You can test if an object is a \code{\link[=query]{query()}} object using \code{\link[=is_query]{is_query()}}.
\code{\link[=to_count_filter]{to_count_filter()}} will allow you to convert a \code{\link[=query]{query()}} to
a vector of filters appropriate for use on \code{\link[=count_mentions]{count_mentions()}}.
}
