% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_read_account.R
\name{read_account}
\alias{read_account}
\title{Read account information}
\usage{
read_account(code)
}
\arguments{
\item{code}{The account code}
}
\value{
The raw json data for an account
}
\description{
Read account information
}
\author{
Constance Neeser
}
