% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_api.R
\name{read_api}
\alias{read_api}
\title{Read arbitrary data from our api.}
\usage{
read_api(endpoint, query = list())
}
\arguments{
\item{endpoint}{The endpoint that you would like to access}

\item{query}{An httr query list}
}
\value{
the json content
}
\description{
This reads arbitrary data from our api, including mention
data.
}
\author{
Constance Neeser
}
