% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_mentions.R
\name{read_data}
\alias{read_data}
\title{Read mention data from the API.}
\usage{
read_data(x, query)
}
\arguments{
\item{x}{An account object.}

\item{query}{A list of query parameters}
}
\value{
A list of data from the API.
}
\description{
Given an account and a query, this reads data from grouse,
or, if test data exists, returns the test data.
}
\details{
Test data should be in the format \code{test01aa_mentions}.
}
