% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_read_mash.R
\name{read_mash}
\alias{read_mash}
\title{Read arbitrary data from mash.}
\usage{
read_mash(endpoint, query = list())
}
\arguments{
\item{endpoint}{The endpoint that you would like to access}

\item{query}{An httr query list}
}
\value{
the json content
}
\description{
This reads data from mash, and performs authentication validation
and so on.
}
\author{
Constance Neeser
}
