% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_rules.R
\name{rules}
\alias{rules}
\alias{rules.brandseyer2.account}
\alias{rules.list}
\alias{rules.data.frame}
\alias{rules.character}
\title{Read account rules}
\usage{
rules(x, .show.progress)

\method{rules}{brandseyer2.account}(x, .show.progress = interactive())

\method{rules}{list}(x, .show.progress = interactive())

\method{rules}{data.frame}(x, .show.progress = interactive())

\method{rules}{character}(x, .show.progress = interactive())
}
\arguments{
\item{x}{An object to read logs from.}

\item{.show.progress}{A logical indicating whether to show a progress bar or not.
By default, this only shows a progress bar in interactive environments.
Also, progress bars will only be shown if 5 or more accounts are
being examined.}
}
\value{
A tibble of rule information,.
}
\description{
Returns a tibble of all the rules for an account. Rules are
applied to mentions as they are received by BrandsEye.
}
\details{
Rules can be read from various different kinds of object. An \code{\link[=account]{account()}}
object is the most basic, although rules can also be read from a list of
account objects (in which case the returned tibble will include an account code column),
or from a tibble containing an account code column, such as that returned
by \code{\link[=account_list]{account_list()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: For an \code{\link[=account]{account()}} object.

\item \code{list}: For a list of account objects.

\item \code{data.frame}: For a \code{data.frame} that contains an \code{account} column of account codes.

\item \code{character}: For a character vector
}}

\examples{
\dontrun{
account("TEST01AA") \%>\%
  rules()
}
\dontrun{
account("TEST01AA", "TEST02AA") \%>\%
  rules()
}
\dontrun{
account_list() \%>\%
  rules()
}

\dontrun{
rules("TEST01AA")
rules(c("TEST01AA", "TEST02AA"))
}
}
