% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_count_filter.R
\name{to_count_filter}
\alias{to_count_filter}
\title{Converts a query to a vector of filters for counting by.}
\usage{
to_count_filter(query, code)
}
\arguments{
\item{query}{A \code{\link[=query]{query()}} object.}

\item{code}{The account we want filters for.}
}
\value{
A vector of characters.
}
\description{
\code{\link[=to_count_filter]{to_count_filter()}} converts a \code{\link[=query]{query()}} to a vector
of filters appropriate for counting by. An account code
is specified to choose a filter for a specific account.
Each filter is required to perform all of the count operations
being requested by the query for an individual account.
}
\examples{

account("TEST01AA") \%>\%
  filter_mentions("published inthelast week") \%>\%
  to_count_filter("TEST01AA")
}
