% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_query.R
\name{to_query}
\alias{to_query}
\title{Create a query object from another object}
\usage{
to_query(x)
}
\arguments{
\item{x}{An object to turn in to a \code{\link[=query]{query()}} object.}
}
\value{
A query object
}
\description{
Creates a \code{\link[=query]{query()}} object from another object. This
can be a convenient way to create queries.
}
\examples{

to_query(account("TEST01AA"))
}
