% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_topic_trees.R
\name{topic_trees}
\alias{topic_trees}
\alias{topic_trees.brandseyer2.account}
\alias{topic_trees.list}
\title{List topic trees for an account.}
\usage{
topic_trees(x, ...)

\method{topic_trees}{brandseyer2.account}(x, ...)

\method{topic_trees}{list}(x, ...)
}
\arguments{
\item{x}{A account or list of accounts.}

\item{...}{Arguments for other methods.}
}
\description{
Returns a tibble of topic trees associated with the account.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Provides topic trees for \code{\link[=account]{account()}} objects.

\item \code{list}: Get topic trees for a list of accounts.
}}

\examples{

account("TEST01AA") \%>\%
  topic_trees()

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  topic_trees()
}
