% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_update_mentions.R
\name{update_mentions}
\alias{update_mentions}
\alias{update_mentions.brandseyer2.account.v4}
\title{Update data stored on a mention}
\usage{
update_mentions(x, filter, auto.confirm = FALSE, ...)

\method{update_mentions}{brandseyer2.account.v4}(
  x,
  filter,
  auto.confirm,
  ...,
  relevancy,
  relevancyVerified,
  addTags,
  removeTags,
  sentiment,
  sentimentVerified,
  media,
  gender,
  language,
  race,
  addPhrases,
  removePhrases,
  addCrowdJobs,
  removeCrowdJobs,
  updateAuthor,
  location
)
}
\arguments{
\item{x}{The account object.}

\item{filter}{Defines the mentions to update.}

\item{auto.confirm}{Some updates, such as updating sentiment, can badly effect an account.
This function will usually ask if you want to proceed. Setting this to
TRUE will turn this behaviour off.}

\item{...}{Arguments for other methods.}

\item{relevancy}{A string indicating whether mentions are relevant or not.}

\item{relevancyVerified}{A boolean indicating whether to consider the relevancy as verifed or not.}

\item{addTags}{A list of tag IDs to add to the mention.}

\item{removeTags}{A list of tag IDs to remove.}

\item{sentiment}{An integer indicating sentiment.}

\item{sentimentVerified}{A boolean indicating whether to consider sentiment as verified or not.}

\item{media}{A string indicating the category of the mentions, such as CONSUMER or PRESS}

\item{gender}{A string indicating the author's gender, such as MALE, FEMALE, OTHER, UNKNOWN.}

\item{language}{A two letter, lowercase string, giving the iso 639-1 language code for the mention.}

\item{race}{A string giving the race of the author}

\item{addPhrases}{A list of integer IDs for phrases to add to the mention}

\item{removePhrases}{A list of integer IDs for the phrases to remove from the mention}

\item{addCrowdJobs}{A list of integer IDs representing crowd jobs related to this mention, to be
added to the mention.}

\item{removeCrowdJobs}{A list of integer IDs representing crowd jobs to be removed from this mention.}

\item{updateAuthor}{Whether the stored author information related to this mention should be
updated with any of the information related to the author on this mention, such as race,
gender, language or media.}

\item{location}{A string giving location information.}
}
\value{
Nothing
}
\description{
This function lets you update data that is stored on a mention.
For instance, you can add and remove tags from mentions. See \code{\link[=create_tags]{create_tags()}}
to add new tags to an account before adding them to mentions.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account.v4}: 
}}

\note{
Setting \code{sentiment} and \code{relevancy} will affect crowd data. You should strongly think about
setting \code{relevancyVerified} or \code{sentimentVerified} as appropriate, if you decide to edit
\code{sentiment} or \code{relevancy}.
}
\author{
Constance Neeser
}
