% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_users.R
\name{users}
\alias{users}
\alias{users.brandseyer2.account}
\alias{users.character}
\alias{users.list}
\alias{users.data.frame}
\title{Get users for an account}
\usage{
users(x)

\method{users}{brandseyer2.account}(x)

\method{users}{character}(x)

\method{users}{list}(x)

\method{users}{data.frame}(x)
}
\arguments{
\item{x}{An object to get users for, such as an account object.}
}
\value{
A tibble of users.
}
\description{
Returns a tibble of users for an account, or list of accounts.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: List users for an account object.

\item \code{character}: List users for accounts given as account codes.

\item \code{list}: List users for a list of account objects.

\item \code{data.frame}: List users for accounts coded in the \code{account} column of a tibble, such as
provided by \code{\link[=account_list]{account_list()}}.
}}

\examples{
\dontrun{
account("TEST01AA") \%>\%
  users()
}

\dontrun{
users("TEST01AA")
users(c("TEST01AA", "TEST02AA"))
}
\dontrun{
account("TEST01AA", "TEST02AA") \%>\%
  users()
}
\dontrun{
account_list() \%>\%
  users()
}
}
