% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{whoami}
\alias{whoami}
\title{Returns your current authentication credentials.}
\usage{
whoami(raise_error = FALSE)
}
\arguments{
\item{raise_error}{If this is set to \code{TRUE}, the function
will raise an error instead of returning NULL if you
are not authenticated.}
}
\value{
Returns \code{NULL} if you are not logged in,
otherwise returns your current authentication credentials.
}
\description{
If you need to authenticate yourself, you can do this
with the \code{\link[=authenticate]{authenticate()}} function. \code{\link[=am_i_brandseye]{am_i_brandseye()}}
is a useful method for determining if you're a staff member
or not. The \code{raise_error} method makes \code{\link[=whoami]{whoami()}} an easy
way to determine if the user is authenticated, and to raise
an error if not.
}
\author{
Constance Neeser
}
