% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_language.R
\name{with_fields}
\alias{with_fields}
\title{Select extra data when fetching or counting mentions}
\usage{
with_fields(.account, ..., .envir)
}
\arguments{
\item{.account}{An account or \code{\link[=query]{query()}} object.}

\item{...}{The fields to order by.}

\item{.envir}{An environment to perform substitutions in.}
}
\value{
A \code{\link[=query]{query()}} object
}
\description{
\code{\link[=with_fields]{with_fields()}} allows you to fetch additional data fields on mentions,
or to count additional fields when counting mentions. This is part of the
\code{\link[=query]{query()}} language.
}
\examples{

account("TEST01AA") \%>\%
  filter_mentions("published inthelast week") \%>\%
  with_fields(mentionCount, totalOTS)
}
\seealso{
Other verbs for the query language: \code{\link[=group_mentions_by]{group_mentions_by()}},
\code{\link[=compare_mentions]{compare_mentions()}}, \code{\link[=filter_mentions]{filter_mentions()}}, \code{\link[=with_mention_order]{with_mention_order()}}, \code{\link[=with_account]{with_account()}}.

\code{\link[=query]{query()}} is a way to manually create queries.
}
