% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_with_tag_parents.R
\name{with_tag_parents}
\alias{with_tag_parents}
\title{Determine the parent of tags given a particular tree.}
\usage{
with_tag_parents(tags, parent_id)
}
\arguments{
\item{tags}{A tibble of tags to find parents for}

\item{parent_id}{A root parent that all children are determined from}
}
\value{
The same tags tibble, but annotated with a \code{parent} column
}
\description{
Adds a column called \code{parent} to a table. This is the ID
of a tag's parent tag. Since tags can have multiple parents,
we need to know what the ultimate root parent to determine
"parenthood" is. This is done using the \code{parent_id} parameter.
}
\examples{
# Add parents to tags
account("TEST01AA") \%>\%
  tags() \%>\%
  with_tag_parents(10)

# See what root brands we have
account("TEST01AA") \%>\%
  topic_trees()

# Tag 1001 is a topic tree. Let's find
# topics, with parents, that are part of this tree.
account("TEST01AA") \%>\%
  tags() \%>\%
  with_tag_parents(1001) \%>\%
  dplyr::filter(namespace == 'topic') \%>\%
  dplyr::arrange(desc(is_parent), name)
}
