% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_write_api.R
\name{write_api}
\alias{write_api}
\title{Write arbitrary data to our api.}
\usage{
write_api(endpoint, method = "PUT", json = NULL, query = list())
}
\arguments{
\item{endpoint}{The endpoint that you would like to access}

\item{method}{A string indicating the kind of write action. Possibly 'PUT' or 'POST'}

\item{json}{An object to serialise as json.}

\item{query}{An httr query list}
}
\value{
the json content
}
\description{
This reads arbitrary data from our api, including mention
data.
}
\author{
Constance Neeser
}
