% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_write_mash.R
\name{write_mash}
\alias{write_mash}
\title{Write arbitrary data to mash.}
\usage{
write_mash(endpoint, method = "PUT", json = NULL, query = list())
}
\arguments{
\item{endpoint}{The endpoint that you would like to access}

\item{method}{A string indicating the kind of write action. Possibly 'PUT' or 'POST'}

\item{json}{An object to serialise as json.}

\item{query}{An httr query list}
}
\value{
the json content
}
\description{
This reads arbitrary data from mash.
}
\author{
Constance Neeser
}
