% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account.R
\name{account}
\alias{account}
\alias{accounts}
\alias{account.character}
\alias{account.brandseyer2.account_list}
\title{Access account information}
\usage{
account(code)

\method{account}{character}(code)

\method{account}{brandseyer2.account_list}(codes)
}
\arguments{
\item{code}{A vector of one or more account codes.}

\item{codes}{A tibble provided by \code{\link{account_list}}

Get account information from the list of accounts available to from \code{\link{account_list}}}
}
\value{
An account object if one account code is given, or a list of account objects,
        one for each code given, in the same order as the codes were given.
}
\description{
Access account information
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Fetch counts for one or more account codes given as a character vector.

\item \code{brandseyer2.account_list}: 
}}

\examples{

\dontrun{

# Read one account
account("TEST01AA")

# Read two accounts, returned as a list.
accounts(c("TEST01AA", "TEST02AA"))
}
}
\seealso{
\code{\link{accounts}} is a synonym for \code{account}.
}
