% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account.R
\name{account}
\alias{account}
\alias{accounts}
\alias{account.character}
\alias{account.brandseyer2.account_list}
\title{Access account information}
\usage{
account(codes, ...)

\method{account}{character}(codes, ...)

\method{account}{brandseyer2.account_list}(codes, ...)
}
\arguments{
\item{codes}{A vector of one or more account codes, or possibly
a tibble, as provided by \code{\link[=account_list]{account_list()}}}

\item{...}{Additional account codes.}
}
\value{
An account object if one account code is given, or a list of account objects,
one for each code given, in the same order as the codes were given.
}
\description{
Returns an account object, representing the account that you're
performing actions on in brandseyer2.
}
\details{
Typically you may ask for only a single account, by passing
a single account code as an argument. You can also pass in multiple
account codes, or even a vector of account codes, to be given a list
of account objects to act on.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Fetch counts for one or more account codes given as a character vector.

\item \code{brandseyer2.account_list}: Get account information from the list of accounts available to from \code{\link[=account_list]{account_list()}}
}}

\examples{

# Read one account
account("TEST01AA")

# Read two accounts
account("TEST01AA", "TEST02AA")

# Read accounts from a vector
account(c("TEST01AA", "TEST02AA"))
}
\seealso{
\code{\link[=accounts]{accounts()}} is a synonym for \code{account}.

\code{\link[=is.account]{is.account()}} to test if an object is an account object.

\code{\link[=tags]{tags()}} to find tags in an account.

\code{\link[=topics]{topics()}} to find topics in an account.

\code{\link[=create_tags]{create_tags()}} to add more tags to an account.

\code{\link[=topic_trees]{topic_trees()}} to find the topic trees in an account.
}
