% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account.R
\name{account}
\alias{account}
\title{Access account information}
\usage{
account(code)
}
\arguments{
\item{code}{The account code}
}
\value{
An account object.
}
\description{
Access account information
}
