% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account.R
\name{account}
\alias{account}
\title{Access account information}
\usage{
account(code)
}
\arguments{
\item{codes}{A vector of one or more account codes.}
}
\value{
An account object if one account code is given, or a list of account objects,
        one for each code given, in the same order as the codes were given.
}
\description{
Access account information
}
\examples{

\dontrun{

# Read one account
account("TEST01AA")

# Read two accounts, returned as a list.
account(c("TEST01AA", "TEST02AA"))
}
}
