% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account.R
\name{account}
\alias{account}
\alias{accounts}
\alias{account.character}
\alias{account.data.frame}
\title{Access account information}
\usage{
account(codes, ...)

\method{account}{character}(codes, ..., .show.progress = interactive())

\method{account}{data.frame}(codes, ..., .show.progress = interactive())
}
\arguments{
\item{codes}{A vector of one or more account codes, or possibly
a tibble (with an \code{account} column, such as from \code{\link[=account_list]{account_list()}})}

\item{...}{Additional account codes.}

\item{.show.progress}{A logical indicating whether a progress bar should
be shown or not. By default, it will only be shown on interactive sessions.
Further, no matter this parameter's value, the bar will only be shown
when fetching account information for more than five accounts.}
}
\value{
An account object if one account code is given, or a list of account objects,
one for each code given, in the same order as the codes were given.
}
\description{
Returns an account object, representing the account that you're
performing actions on in brandseyer2.
}
\details{
Typically you may ask for only a single account, by passing
a single account code as an argument. You can also pass in multiple
account codes, or even a vector of account codes, to be given a list
of account objects to act on.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Fetch counts for one or more account codes given as a character vector.

\item \code{data.frame}: Get account information from a data.frame. The data.frame
should have an \code{account} column holding the account's code.
\code{\link[=account_list]{account_list()}} is a function that returns just such a data.frame.
}}

\section{Interactive use}{


When reading multiple accounts at a time, this method will by defualt
show a progress bar in interactive environments (such as RStudio). You
can control this behaviour using the \code{.show.progress} parameter.
}

\examples{

# Read one account
account("TEST01AA")

# Read two accounts
account("TEST01AA", "TEST02AA")

# Read accounts from a vector
account(c("TEST01AA", "TEST02AA"))
}
\seealso{
\code{\link[=accounts]{accounts()}} is a synonym for \code{account}.

\code{\link[=is_account]{is_account()}} to test if an object is an account object.

\code{\link[=tags]{tags()}} to find tags in an account.

\code{\link[=topics]{topics()}} to find topics in an account.

\code{\link[=create_tags]{create_tags()}} to add more tags to an account.

\code{\link[=topic_trees]{topic_trees()}} to find the topic trees in an account.

\code{\link[=logs]{logs()}} to read the account's logs.

\code{\link[=users]{users()}} to see what users have access to the account.
}
