% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_code.R
\name{account_code}
\alias{account_code}
\title{Get an account's code}
\usage{
account_code(account)
}
\arguments{
\item{account}{The account object.}
}
\value{
A string holding the account's code
}
\description{
Get an account's code
}
