% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_list.R
\name{account_list}
\alias{account_list}
\title{List all accounts available to you.}
\usage{
account_list()
}
\value{
A tibble of accounts available to you.
}
\description{
These accounts are provided as a tibble. The tibble can be
passed to some other functions, such as \code{\link[=account]{account()}},
to return information from across all accounts available to you.
}
\examples{

\dontrun{
# Fetch the tags from all accounts available to you.
account_list() \%>\%
  tags()
}
}
