% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_tags.R
\name{account_tags}
\alias{account_tags}
\alias{account_tags.brandseyer2.account}
\alias{account_tags.list}
\title{Fetch tags for an account}
\usage{
account_tags(accounts)

\method{account_tags}{brandseyer2.account}(accounts)

\method{account_tags}{list}(accounts)
}
\arguments{
\item{accounts}{One or more account objects.}
}
\value{
A tibble of tag information. Includes the name, namespace, and description of the
tags. Note that topics are stored in the 'topics' namespace.
}
\description{
Fetches tag information information, returned as a tibble, for the given
account.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Get tags from an account object.

\item \code{list}: Create a table of tags for the list of accounts given
}}

\examples{
\dontrun{

# See what namespaces are in your account
account("TEST01AA") \%>\%
  account_tags() \%>\%
  dplyr::select(namespace) \%>\%
  table()

# Find topics
account("TEST01AA") \%>\%
  account_tags() \%>\%
  dplyr::filter(namespace == "topic")
}


\dontrun{
accounts(c("TEST01AA", "TEST02AA")) \%>\%
  account_tags()
}
}
\seealso{
\code{\link{account_topics}} to see just a list of topics.
}
