% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_account_topics.R
\name{account_topics}
\alias{account_topics}
\title{Fetch topics for an account}
\usage{
account_topics(account)
}
\arguments{
\item{account}{An account object.}
}
\value{
A tibble of topic information. Includes the name and topic description.
}
\description{
Fetches the topics that are used in an account.
}
