% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_brands.R, R/brands.R,
%   R/mention_brands.R
\name{brands.brandseyer2.account}
\alias{brands.brandseyer2.account}
\alias{brands.list}
\alias{brands}
\alias{brands.data.frame}
\title{Fetch brand information}
\usage{
\method{brands}{brandseyer2.account}(x, ...)

\method{brands}{list}(x, ...)

brands(x, ...)

\method{brands}{data.frame}(x, ..., ac = attr(x, "account"))
}
\arguments{
\item{x}{An object to fetch brand data for, such as an \code{\link{account}}
or tibble of mention data.}

\item{...}{Additional parameters for functions.}

\item{ac}{An optional account object from which to take tag information.}
}
\value{
A tibble of brand information
}
\description{
Fetches brand information for accounts and for mention data.
This is returned returned as a tibble. For account objects,
this returns the brands that make up the account. For mention data,
this returns the brands that are associated with the mention.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Read brands for only a single account

\item \code{list}: Create a table of brands for the list of x given

\item \code{data.frame}: Fetch brand information from a tibble of mention data.
}}

\examples{

# Fetch brand information
account("TEST01AA") \%>\%
  brands()

# Find deleted parent brands
account("TEST01AA") \%>\%
  brands() \%>\%
  dplyr::filter(is.na(parent), deleted)

# Fetch phrases without using `phrases``
account("TEST01AA") \%>\%
  brands() \%>\%
  dplyr::select(id, phrases) \%>\%
  dplyr::rename(brand.id = id) \%>\%
  tidyr::unnest(phrases) \%>\%
  dplyr::rename(phrase.id = id)

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  brands()
}
\seealso{
\code{\link{root_brands}} for filtering to only the root brands of an account.
}
\author{
Constance Neeser
}
