% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_brands.R
\name{brands}
\alias{brands}
\alias{brands.brandseyer2.account}
\alias{brands.list}
\title{Fetch account brands}
\usage{
brands(accounts)

\method{brands}{brandseyer2.account}(accounts)

\method{brands}{list}(accounts)
}
\arguments{
\item{accounts}{One or more account objects.}
}
\value{
A tibble of brand information
}
\description{
Fetches brand information, returned as a tibble, for the given
account. It includes brand names, associated phrases, and so on.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Read brands for only a single account

\item \code{list}: Create a table of brands for the list of accounts given
}}

\examples{

# Fetch brand information
account("TEST01AA") \%>\%
  brands()

# Find deleted parent brands
account("TEST01AA") \%>\%
  brands() \%>\%
  dplyr::filter(is.na(parent), deleted)

# Fetch phrases without using `phrases``
account("TEST01AA") \%>\%
  brands() \%>\%
  dplyr::select(id, phrases) \%>\%
  dplyr::rename(brand.id = id) \%>\%
  tidyr::unnest(phrases) \%>\%
  dplyr::rename(phrase.id = id)

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  brands()
}
\seealso{
\code{\link{root_brands}} for filtering to only the root brands of an account.
}
\author{
Constance Neeser
}
