% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_language.R
\name{compare_mentions}
\alias{compare_mentions}
\title{Compare mentions using sub-filters when counting them}
\usage{
compare_mentions(.account, ...)
}
\arguments{
\item{.account}{An account or \code{\link[=query]{query()}} object.}

\item{...}{A named set of sub-filters, such as
\code{positive = sentiment = 1}. See the examples.}
}
\value{
A \code{\link[=query]{query()}} object.
}
\description{
\code{\link[=compare_mentions]{compare_mentions()}} lets you group mentions together using
sub-filters when counting them using \code{\link[=count_mentions]{count_mentions()}}. This
is part of the \code{\link[=query]{query()}} language.
}
\examples{

account("TEST01AA") \%>\%
  compare_mentions(enterprise = "media is enterprise",
                   consume = "media is consumer")
}
\seealso{
Other verbs for the query language: \code{\link[=group_mentions_by]{group_mentions_by()}},
\code{\link[=filter_mentions]{filter_mentions()}}, \code{\link[=with_mention_fields]{with_mention_fields()}}, \code{\link[=with_mention_order]{with_mention_order()}}, \code{\link[=with_account]{with_account()}}.

\code{\link[=query]{query()}} is a way to manually create queries.
}
