% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_count.R
\name{count_mentions}
\alias{count_mentions}
\title{Count mentions}
\usage{
count_mentions(x, filter, groupBy, ...)
}
\arguments{
\item{x}{An account to read from}

\item{filter}{A filter to use}

\item{groupBy}{An optional list of names of things to group by.
For example, \code{groupBy = published}, or,
\code{groupBy = tag}}

\item{...}{Further arguments for other methods}
}
\value{
A tibble of data.
}
\description{
This counts mentions, aggregating them in to various data sets.
}
\examples{

\dontrun{

# Count all mentions in your account in the last year
account("TEST01AA") \%>\% 
  count_mentions("published inthelast year and brandisorchildof 1")
  
# Count all mentions in your account, grouping by publication date.
account("TEST01AA") \%>\% 
  count_mentions("published inthelast year and brandisorchildof 1", groupBy = published)

}
}
