% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_count.R
\name{count_mentions}
\alias{count_mentions}
\alias{count_mentions.brandseyer2.account.v4}
\alias{count_mentions.character}
\alias{count_mentions.brandseyer2.query}
\title{Count mentions}
\usage{
count_mentions(.account, ...)

\method{count_mentions}{brandseyer2.account.v4}(
  .account,
  filter,
  ...,
  groupBy = NULL,
  select = NULL,
  orderBy = NULL,
  tagNamespace = NULL,
  .envir = parent.frame()
)

\method{count_mentions}{character}(
  .account,
  filter,
  ...,
  timezone = "Africa/Johannesburg",
  groupBy = NULL,
  select = NULL,
  orderBy = NULL,
  tagNamespace = NULL
)

\method{count_mentions}{brandseyer2.query}(
  .account,
  ...,
  tagNamespace = NULL,
  .show.progress = interactive()
)
}
\arguments{
\item{.account}{An account to read from}

\item{...}{Further arguments for other methods}

\item{filter}{A filter to use}

\item{groupBy}{An optional list of names of things to group by.
For example, \code{groupBy = published}, or,
\code{groupBy = tag}}

\item{select}{An optional list of names of things to select and
aggregate by, in addition to the number of mentions.
For example, \code{select = totalSentiment}, or,
\code{select = c(totalSentiment, totalEngagement)}.}

\item{orderBy}{An optional list of names of the aggregate fields to
order the results by.}

\item{tagNamespace}{An optional string. When grouping by \code{tag}, \code{tagNamespace} can
be supplied to limit the tags being grouped by to only those
in the given namespace. For example, to only see topics,
have \code{tagNamespace = 'topic'}}

\item{.envir}{An optional environment in which to evaluate variables in
your \code{groupBy}, \code{select}, and \code{orderBy} arguments.}

\item{timezone}{A character vector indicating the timezone that dates are
reported in. This should usually be the account's timezone.}

\item{.show.progress}{A logical value to indicate whether to show
a progress bar or not. Defaults to showing the progress bar in
interactive settings. In addition, the progress bar will only be
shown when counting mentions for more than three accounts at a time.}
}
\value{
A tibble of data.
}
\description{
Counts mentions, aggregating them in to various data sets.
See the \code{\link[=query]{query()}} language to conveniently specify what to count.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account.v4}: Count mentions in a v4 account.

\item \code{character}: Allows you to fetch mentions using raw fields. All fields are dynamically
created (such as groupBy) are, in this version of the function, character
vectors.

\item \code{brandseyer2.query}: Count mentions using a \code{\link[=query]{query()}}.
}}

\examples{

\dontrun{

# Count all mentions in your account in the last year
account("TEST01AA") \%>\%
  count_mentions("published inthelast year and brandisorchildof 1")

# Count all mentions in your account, grouping by publication date.
account("TEST01AA") \%>\%
  count_mentions("published inthelast year and brandisorchildof 1", groupBy = published)

}

\dontrun{
  count_mentions("TEST01AA", "published inthelast week", groupBy = c("published"))
}


\dontrun{
account("TEST01AA", "TEST02AA") \%>\%
  filter_mentions("published inthelast week") \%>\%
  group_mentions_by(published) \%>\%
  count_mentions()
}
}
