% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_dashboards.R
\name{dashboards}
\alias{dashboards}
\title{List the dashboards in an account}
\usage{
dashboards(x, d)
}
\arguments{
\item{x}{An account object to query.}

\item{d}{An optional filter}
}
\value{
A tibble of dashboard data.
}
\description{
In \href{https://analyse.brandseye.com}{Analyse}, BrandsEye's analysis tool,
people can set up dashboards to query our data. This function exposese those
dashboards as a tibble, allowing you to query sections, metrics and their
associated filters.
}
\examples{

\dontrun{

# Get all dashboards.
account("TEST01AA") \%>\%
  dashboards()

# Get the dashboard with ID 1.
account("TEST01AA") \%>\%
  dashboards(1)

# Get dashboards 1 and 21.
account("TEST01AA") \%>\%
  dashboards(c(1, 21))

# Get dashboards whose name contains the string 'overview'.
account("TEST01AA") \%>\%
  dashboards("overview")

# Get dashboards whose name contains the string 'overview' or the string 'who'.
account("TEST01AA") \%>\%
  dashboards(c("overview", "who"))

}
}
