% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_language.R
\name{filter_mentions}
\alias{filter_mentions}
\title{Filter for only specific mentions}
\usage{
filter_mentions(.account, filter)
}
\arguments{
\item{.account}{An account or \code{\link[=query]{query()}} object.}

\item{filter}{A filter string.}
}
\value{
A query object
}
\description{
\code{filter_mentions()} causes counting and mention operations
to occur on only those mentions matching the filter. This is
part of the \code{\link[=query]{query()}} language.
}
\examples{

account("TEST01AA") \%>\%
  filter_mentions("published inthelast week")
}
\seealso{
Other verbs for the query language: \code{\link[=group_mentions_by]{group_mentions_by()}},
\code{\link[=compare_mentions]{compare_mentions()}}, \code{\link[=with_fields]{with_fields()}}, \code{\link[=with_order]{with_order()}}

\code{\link[=query]{query()}} is a way to manually create queries.
}
