% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_logs.R
\name{logs}
\alias{logs}
\alias{logs.character}
\alias{logs.brandseyer2.account}
\alias{logs.list}
\alias{logs.data.frame}
\title{Read account logs}
\usage{
logs(x, from, .show.progress)

\method{logs}{character}(x, from = (Sys.Date() \%m-\% months(1)), .show.progress = interactive())

\method{logs}{brandseyer2.account}(x, from = (Sys.Date() \%m-\% months(1)), .show.progress = interactive())

\method{logs}{list}(x, from = (Sys.Date() \%m-\% months(1)), .show.progress = interactive())

\method{logs}{data.frame}(x, from = (Sys.Date() \%m-\% months(1)), .show.progress = interactive())
}
\arguments{
\item{x}{An object to read logs from.}

\item{from}{A date indicating from when to read dates.}

\item{.show.progress}{A logical indicating whether to show a progress bar or not.
By default, this only shows a progress bar in interactive environments.
Also, progress bars will only be shown if 3 or more accounts are
being examined.}
}
\value{
A tibble of log information, including the user that perform the action, and how many
times the action was performed.
}
\description{
Returns a tibble of logs for the account, in the last period stated.
}
\details{
Logs can be read from various different kinds of object. An \code{\link[=account]{account()}}
object is the most basic, although logs can also be read from a list of
account objects (in which case the returned tibble will include an account code column),
or from a tibble containing an account code column, such as that returned
by \code{\link[=account_list]{account_list()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Reach for an account code character vector.

\item \code{brandseyer2.account}: For \code{\link[=account]{account()}} objects.

\item \code{list}: For a list of accounts

\item \code{data.frame}: Read for a data frame, such as returned by \code{\link[=account_list]{account_list()}}
}}

\section{Useful logging functions}{

\itemize{
\item \code{\link[=logs_retrosent]{logs_retrosent()}} filters log information to find retrosend specific information.
}
}

\examples{

\dontrun{
# Defaults to the last month of data
logs("TEST01AA")

# The last week
logs("TEST01AA", Sys.Date() - lubridate::weeks(1))

# The last day
logs("TEST01AA", Sys.Date() - lubridate::days(1))

# The last three months
library(lubridate)
logs("TEST01AA", Sys.Date() \%m-\% months(1))
}


\dontrun{
# Read logs for an account
account("TEST01AA") \%>\%
  logs()
}


\dontrun{
# Read for a list of accounts
account("TEST01AA", "TEST02AA") \%>\%
  logs()
}
\dontrun{
account_list() \%>\%
  logs()
}
}
