% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_mentions.R
\name{mentions}
\alias{mentions}
\alias{mentions.brandseyer2.account.v4}
\alias{mentions.brandseyer2.account.v3}
\title{Fetch mentions from an account}
\usage{
mentions(x, filter, select, ...)

\method{mentions}{brandseyer2.account.v4}(x, filter, select = NULL, ...,
  orderBy = NULL, fetchGraph = FALSE)

\method{mentions}{brandseyer2.account.v3}(x, filter, select, ...,
  use.brandseyer = FALSE, limit = 30, offset = 0, include, all = FALSE)
}
\arguments{
\item{x}{An account object}

\item{filter}{A query to match mentions against. Always required.
See the filter vignette for details.}

\item{select}{A character vector of the mention fields to be returned.}

\item{...}{Further arguments passed to or from other methods.}

\item{orderBy}{Fields to order the returned data by. Defaults to published}

\item{fetchGraph}{Fetch other mentions that are part of the same conversation as this one.}

\item{use.brandseyer}{Set to TRUE if the old brandseyer library is installed,
and you would like to read from an account still using the old V3 API.}

\item{limit}{The maximum number of mentions to be returned}

\item{offset}{Mentions are returned in an order. Offset says how many of the
first mentions should be skipped.}

\item{include}{A character vector of extra information to include in the mentions}

\item{all}{Set to true if you would like to return all mentions from the account.
This overides the \code{limit} parameter.}
}
\value{
A tibble of mentions.
}
\description{
This will give you access to mentions in an account.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account.v4}: Reads V4 accounts. Returns a tibble of mentions.

\item \code{brandseyer2.account.v3}: Reads older, V3 accounts. To use this function, please ensure that
the older \code{brandseyer} library is installed as well. Returns
old \code{brandseyer} data structures, as though the \code{brandseyer::account_mentions}
function had been called.
}}

\note{
It is possible to read mentions from older accounts still
using the V3 API. Using the older brandseyer library is the most
convenient way of doing this, but if the brandseyer library is installed,
the \code{mentions} function will automatically fall back to the appropriate
function calls in \code{brandseyer} to read mentions for you.
}
\examples{

# Reading data from a single account
account("TEST01AA") \%>\%
  mentions(filter = "published inthelast week and brand isorchildof 1") # Must always have a filter
}
\seealso{
\code{\link{tags}}() to fetch tag information from mentions.

\code{\link{topics}}() to fetch topic information from mentions.

\code{\link{brands}}() to fetch brand information from mentions.

\code{\link{phrases}}() to fetch phrase information from mentions.
}
