% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentions_mentions.R
\name{mentions}
\alias{mentions}
\alias{mentions.brandseyer2.account.v4}
\title{Fetch mentions from an account}
\usage{
mentions(x, filter, select, ..., .envir)

\method{mentions}{brandseyer2.account.v4}(
  x,
  filter,
  select = NULL,
  ...,
  .envir = parent.frame(),
  orderBy = NULL,
  fetchGraph = FALSE
)
}
\arguments{
\item{x}{An account object}

\item{filter}{A query to match mentions against. Always required.
See the filter vignette for details.}

\item{select}{A character vector of the mention fields to be returned.}

\item{...}{Further arguments passed to or from other methods.}

\item{.envir}{An optional environment for when substituting values in to select.}

\item{orderBy}{Fields to order the returned data by. Defaults to published}

\item{fetchGraph}{Fetch other mentions that are part of the same conversation as this one.}
}
\value{
A tibble of mentions. Returns an empty tibble if there are no mentions.
}
\description{
This provides access to an account's mentions.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account.v4}: Read mentions from an account. Returns a tibble of mention data.
}}

\note{
We cannot read data from older, V3 accounts / APIs.
}
\section{Availability of content}{


Tweets, by default, will be given with no content, only
URIs and IDs. This is per our agreement with twitter. Twitter
does provide an API for obtaining tweet content using tweet IDs.

If you fetch mentions and choose not to select the \code{site} field
(selected by default), all mentions may contain no data.
}

\examples{

# Reading data from a single account
account("TEST01AA") \%>\%
  mentions(filter = "published inthelast week and brand isorchildof 1") # Must always have a filter
}
\seealso{
\url{tags()} to fetch tag information from mentions.

\url{topics()} to fetch topic information from mentions.

\url{brands()} to fetch brand information from mentions.

\url{phrases()} to fetch phrase information from mentions.
}
