% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_phrases.R, R/mention_phrases.R,
%   R/phrases.R
\name{phrases.brandseyer2.account}
\alias{phrases.brandseyer2.account}
\alias{phrases.list}
\alias{phrases.data.frame}
\alias{phrases}
\title{Fetch phrases for an accounts and mentions}
\usage{
\method{phrases}{brandseyer2.account}(x, ...)

\method{phrases}{list}(x, ...)

\method{phrases}{data.frame}(x, ..., ac = attr(x, "account"))

phrases(x, ...)
}
\arguments{
\item{x}{An object to read phrases from, such as an \code{\link[=account]{account()}}
object, or a tibble of mentions.}

\item{...}{Additional arguments for other methods.}

\item{ac}{An optional account object from which to take phrase information.}
}
\value{
A tibble of phrase information.
}
\description{
Fetches phrase information, returned as a tibble, for the given
account. All of this information is included in \code{\link[=brands]{brands()}}
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Read phrases for only a single account

\item \code{list}: Create a table of phrases for the list of accounts given

\item \code{data.frame}: Fetch phrase information from a tibble of mention data.
}}

\note{
\code{\link[=mentions]{mentions()}} should be called with \code{select = "phrases"}
to ensure that phrase information is returned with the mentions.
}
\examples{

# Fetch phrases for a single account
account("TEST01AA") \%>\%
  phrases()

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  phrases()
}
\author{
Constance Neeser
}
