% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_phrases.R, R/phrases.R
\name{phrases.brandseyer2.account}
\alias{phrases.brandseyer2.account}
\alias{phrases.list}
\alias{phrases}
\title{Fetch account phrases}
\usage{
\method{phrases}{brandseyer2.account}(x)

\method{phrases}{list}(x)

phrases(x, ...)
}
\arguments{
\item{x}{An object to read phrases from, such as an \code{\link{account}}
object, or a tibble of mentions.}
}
\value{
A tibble of phrase information.
}
\description{
Fetches phrase information, returned as a tibble, for the given
account. All of this information is included in \code{\link{brands}}
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Read phrases for only a single account

\item \code{list}: Create a table of phrases for the list of accounts given
}}

\examples{

# Fetch phrases for a single account
account("TEST01AA") \%>\%
  phrases()

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  phrases()
}
\author{
Constance Neeser
}
