% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_sections.R
\name{sections}
\alias{sections}
\alias{sections.data.frame}
\title{Give section information for a dashboard}
\usage{
sections(x, d, unnest)

\method{sections}{data.frame}(x, d, unnest = FALSE)
}
\arguments{
\item{x}{A tibble to get section information for.}

\item{d}{An optional value to filter the sections on. Can be a vector
of integer IDs for the wanted sections, or a vector of
characters giving words to perform partial string matches
on section titles with.}

\item{unnest}{A logical value indicating whether sections should be
appended as a list of tibbles, or whether they should be unnested
in to the tibble itself, similar to having called \code{\link[tidyr:unnest]{tidyr::unnest()}}.}
}
\value{
The original tibble, but now with an additional section column.
}
\description{
Given a tibble of dashboard information (from, for example, \code{dashboards()}),
this adds the section information to the tibble.
}
\details{
The second argument to this function, \code{d}, lets you search for specific
mentions in a similar way that \code{\link[=dashboards]{dashboards()}} does for dashboards. See
the \code{dashboards()} documentation for more details.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Get section information for a dashboard tibble.
}}

\examples{

\dontrun{

# Get all sections for the dashboard with ID 1
account("TEST01AA") \%>\%
  dashboards(1) \%>\%
  sections()

# Get the section with ID 5 for the dashboard with ID 1
account("TEST01AA") \%>\%
  dashboards(1) \%>\%
  sections(5)

# Get the sections whose title contain the string "where"
account("TEST01AA") \%>\%
  dashboards(1) \%>\%
  sections("where")
}
}
\seealso{
\code{\link[=dashboards]{dashboards()}} to pull dashboard information for an account.

\code{\link[=metrics]{metrics()}} to pull metric information for sections.
}
