% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_topics.R, R/topics.R
\name{topics.brandseyer2.account}
\alias{topics.brandseyer2.account}
\alias{topics.list}
\alias{topics}
\title{Fetch topics for an account or tibble of mentions}
\usage{
\method{topics}{brandseyer2.account}(x)

\method{topics}{list}(x)

topics(x)
}
\arguments{
\item{x}{Objects to return topic data from.}
}
\value{
A tibble of topic information. Includes the name and topic description.
}
\description{
Fetches the topics that are used in an account, or those
that appear on a tibble of mentions.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Get topics for a single account.

\item \code{list}: Get topics for a list of accounts
}}

\examples{

account("TEST01AA") \%>\%
  topics()

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  topics()
}
