% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_topics.R, R/mention_topics.R,
%   R/topics.R
\name{topics.brandseyer2.account}
\alias{topics.brandseyer2.account}
\alias{topics.list}
\alias{topics.data.frame}
\alias{topics}
\title{Fetch topics for an account or tibble of mentions}
\usage{
\method{topics}{brandseyer2.account}(x, ...)

\method{topics}{list}(x, ...)

\method{topics}{data.frame}(x, ..., ac = attr(x, "account"), na.rm = TRUE)

topics(x, ...)
}
\arguments{
\item{x}{Objects to return topic data from.}

\item{...}{Additional arguments for other methods.}

\item{ac}{An optional account object from which to take topic information.}

\item{na.rm}{Whether to remove mentions that have no topics.}
}
\value{
A tibble of topic information. Includes the name and topic description.
}
\description{
Fetches the topics that are used in an account, or those
that appear on a tibble of mentions.
}
\section{Methods (by class)}{
\itemize{
\item \code{brandseyer2.account}: Get topics for a single account.

\item \code{list}: Get topics for a list of accounts

\item \code{data.frame}: Fetch topic information from a tibble of mention data.
}}

\examples{

account("TEST01AA") \%>\%
  topics()

accounts(c("TEST01AA", "TEST02AA")) \%>\%
  topics()
}
