% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{whoami}
\alias{whoami}
\title{Returns your current authentication credentials.}
\usage{
whoami(raise_error = FALSE)
}
\arguments{
\item{raise_error}{If this is set to \code{TRUE}, the function
will raise an error instead of returning NULL if you
are not authenticated.}
}
\value{
Returns \code{NULL} if you are not logged in,
        otherwise returns your current authentication credentials.
}
\description{
Returns your current authentication credentials.
}
\author{
Constance Neeser
}
